+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -13,6 +13,7 @@ group: Deployment & Configs
 - Added support for linked elements on query components.
 - Added 'Query References' component.
 - Added 'Query Namesake Element' component.
+- Now 'Query Elements' give elements in reverse order when 'Limit' is negative.
 
 ### RC